add visualize code

+++ tools/infer/predict_eval.py
@@ -22,28 +22,27 @@ import numpy as np
 import math
 import time
 import json
+import os
+from PIL import Image, ImageDraw, ImageFont
+from tools.infer.utility import draw_ocr
+from ppocr.utils.utility import get_image_file_list
 
 if __name__ == "__main__":
     args = utility.parse_args()
     text_sys = predict_system.TextSystem(args)
 
-    image_file_list = []
-    label_file_path = "./eval_perform/gt_res/test_chinese_ic15_500_4pts.txt"
-    img_set_path = "./eval_perform/"
-    with open(label_file_path, "rb") as fin:
-        lines = fin.readlines()
-        for line in lines:
-            substr = line.decode('utf-8').strip("\n").split("\t")
-            if "lsvt" in substr[0]:
-                continue
-            image_file_list.append(substr[0])
+    if not os.path.exists(args.image_dir):
+        raise Exception("{} not exists !!".format(args.image_dir))
+    image_file_list = get_image_file_list(args.image_dir)
 
     total_time_all = 0
     count = 0
-    save_path = "./output/predict.txt"
+    save_path = "./inference_output/predict.txt"
+    if not os.path.exists(os.path.dirname(save_path)):
+        os.makedirs(os.path.dirname(save_path))
     fout = open(save_path, "wb")
     for image_name in image_file_list:
-        image_file = img_set_path + image_name
+        image_file = image_name
         img = cv2.imread(image_file)
         if img is None:
             logger.info("error in loading image:{}".format(image_file))
@@ -68,6 +67,23 @@ if __name__ == "__main__":
                 "points": points,
                 "scores": score * 1.0
             })
+        # draw predict box and text in image
+        # and save drawed image in save_path
+        image = Image.open(image_file)
+        boxes, txts, scores = [], [], []
+        for dic in bbox_list:
+            boxes.append(dic['points'])
+            txts.append(dic['transcription'])
+            scores.append(round(dic['scores'], 3))
+        new_img = draw_ocr(image, boxes, txts, scores, draw_txt=True)
+        draw_img_save = os.path.join(
+            os.path.dirname(save_path), "inference_draw",
+            os.path.basename(image_file))
+        if not os.path.exists(os.path.dirname(draw_img_save)):
+            os.makedirs(os.path.dirname(draw_img_save))
+        cv2.imwrite(draw_img_save, new_img[:, :, ::-1])
+        print("drawed img saved in {}".format(draw_img_save))
+        # save predicted results in txt file
         otstr = image_name + "\t" + json.dumps(bbox_list) + "\n"
         fout.write(otstr.encode('utf-8'))
     avg_time = total_time_all / count
